+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Better warnings handling.
 Added a warning if you do not zero-fill enough to do a Hilbert transform.
 29/01/26
 New GUIs for integration. Called by the *integrate* methods of *Spectrum_1D* and *Pseudo_2D*.
+02/02/26
+Reading and writing dosy input/output files
 
 ## New functions:
 21/01/26
@@ -14,6 +16,9 @@ New GUIs for integration. Called by the *integrate* methods of *Spectrum_1D* and
 28/01/26
 - *anal.write_igrl*
 - *anal.read_igrl*
+02/02/26
+- *fit.write_dy*
+- *fit.read_dy*
 
 ## Modified functions:
 21/01/26